6 семестр/Курсовая работа по Базам Данных/Project/requirements.txt
python-telegram-bot==13.7
python-docx==0.8.11
docxtpl==0.11.5
psycopg2==2.9.1
pytz==2019.3